requirements.txt
django 
psycopg2-binary
djangorestframework 
django-filter 
pillow 
whitenoise
gunicorn
python-dotenv
django-cors-headers
